+++ lib/td_lm_web/views/relation_view.ex
@@ -4,8 +4,6 @@ defmodule TdLmWeb.RelationView do
 
   alias TdLmWeb.RelationView
 
-  @tag_attrs [:id, :value]
-
   def render("index.json", %{relations: relations, hypermedia: hypermedia}) do
     render_many_hypermedia(relations, hypermedia, RelationView, "relation.json")
   end
@@ -24,23 +22,20 @@ defmodule TdLmWeb.RelationView do
   end
 
   defp relation_json(relation) do
-    %{
-      id: relation.id,
-      context: relation.context,
-      source_id: relation.source_id,
-      source_type: relation.source_type,
-      target_id: relation.target_id,
-      target_type: relation.target_type,
-      tags: parse_relation_tags(relation)
-    }
+    tags = parse_relation_tags(relation)
+
+    relation
+    |> Map.take([:id, :context, :source_id, :source_type, :target_id, :target_type])
+    |> Map.put(:tags, tags)
   end
 
   defp parse_relation_tags(relation) do
     case Ecto.assoc_loaded?(relation.tags) do
       true ->
         relation
-          |> Map.get(:tags, [])
-          |> Enum.map(&Map.take(&1, @tag_attrs))
+        |> Map.get(:tags, [])
+        |> Enum.map(&Map.take(&1, [:id, :value]))
+
       false ->
         []
     end